+++ Wolf1.Core/Accounting/Property.cs
@@ -8,7 +8,7 @@ namespace Wolf1.Core.Accounting
         public IAddress Location { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public IPerson Owner { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public bool Rental { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        public int Owed { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public decimal Owed { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public DateTime MortgageStart { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public DateTime MortgageEnd { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public string AssetTag { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
@@ -16,7 +16,7 @@ namespace Wolf1.Core.Accounting
         public decimal PurchasePrice { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public DateTime PurchaseDate { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public int PaymentsRemaining { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        public int PaymentAmount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public decimal PaymentAmount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public IHistory ServiceHistory { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public decimal Depreciation { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
     }
